requirements.txt
MetaTrader5
pandas
numpy
python-dotenv
pytest
black
flake8
mypy 